ELT-Airflow-Orchestration/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==2.8.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-postgres==5.14.0
apache-airflow-providers-sqlite==3.9.1
apispec==6.8.4
argcomplete==3.6.3
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.2
cramjam==2.11.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
dash==3.3.0
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
email-validator==1.3.1
fastparquet==2024.11.0
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-re2==1.1.20251105
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
limits==5.6.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.7.0
narwhals==2.12.0
nest-asyncio==1.6.0
numpy==1.24.4
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pathlib_abc==0.5.2
pathspec==0.12.1
pendulum==3.1.0
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.4.2
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.2
rpds-py==0.29.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
setproctitle==1.3.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
streamlit==1.51.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.3.6
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==2.2.3
wrapt==2.0.1
WTForms==3.2.1
yarl==1.22.0
zipp==3.23.0
